FileMemoryInterface
